requirements.txt
python-dotenv
pandas
numpy
matplotlib
mplfinance
httpx
jinja2
openpyxl
markdown
requests
